refactor types. refactor src
